Refactor PVCs strategies by extracting common functionality
